requirements.txt
# Packages written by the author
nldk

# Packages trusted with inexact version specified
nltk==3.*
numpy==1.*
pandas==1.*
python-dotenv==1.*
scikit-learn==1.*
torch==2.*

# Need exact version specified (untrusted)
accelerate==0.28.0
datasets==2.14.6
evaluate==0.4.1
regex==2023.10.3
transformers==4.37.1